Main Credential Script Logic for PowerShell v2
PowerShell v2 is the default in Windows Server 2008. The credential saving and importing are different from PowerShell v3. This script includes the basic logic to save the credential, then call it in order to use the username and password in your code. Note when setting up a task scheduler job in this manner, the user running the task scheduler job must be the same user as the one whose credentials are saved.